refactor(schema): Pull RustVersion out into a mod
